% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{get_log_level}
\alias{get_log_level}
\title{Retrieved the current log level (increasing order of severity: \code{\link{LOG_DEBUG}}, \code{\link{LOG_INFO}}, \code{\link{LOG_WARN}}, \code{\link{LOG_ERROR}}).}
\usage{
get_log_level()
}
\value{
the label for the current log level
}
\description{
Retrieved the current log level (increasing order of severity: \code{\link{LOG_DEBUG}}, \code{\link{LOG_INFO}}, \code{\link{LOG_WARN}}, \code{\link{LOG_ERROR}}).
}
