% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{log_debug}
\alias{log_debug}
\alias{DEBUG}
\title{Logs a message with \code{\link{LOG_DEBUG}} severity.}
\usage{
log_debug(text = "DEBUG message", prefix = NA)

DEBUG(text = "DEBUG message", prefix = NA)
}
\arguments{
\item{text}{The message to log}

\item{prefix}{A message-specific prefix (defaults to \code{NA})}
}
\description{
Logs a message with \code{\link{LOG_DEBUG}} severity.
}
