% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{log_error}
\alias{log_error}
\alias{ERROR}
\title{Logs a message with \code{\link{LOG_ERROR}} severity.}
\usage{
log_error(text = "ERROR message", prefix = NA)

ERROR(text = "ERROR message", prefix = NA)
}
\arguments{
\item{text}{The message to log}

\item{prefix}{A message-specific prefix (defaults to \code{NA})}
}
\description{
Logs a message with \code{\link{LOG_ERROR}} severity.
}
