% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{log_info}
\alias{log_info}
\alias{INFO}
\title{Logs a message with \code{\link{LOG_INFO}} severity.}
\usage{
log_info(text = "INFO message", prefix = NA)

INFO(text = "INFO message", prefix = NA)
}
\arguments{
\item{text}{The message to log}

\item{prefix}{A message-specific prefix (defaults to \code{NA})}
}
\description{
Logs a message with \code{\link{LOG_INFO}} severity.
}
