% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{log_warn}
\alias{log_warn}
\alias{WARN}
\title{Logs a message with \code{\link{LOG_WARN}} severity.}
\usage{
log_warn(text = "WARN message", prefix = NA)

WARN(text = "WARN message", prefix = NA)
}
\arguments{
\item{text}{The message to log}

\item{prefix}{A message-specific prefix (defaults to \code{NA})}
}
\description{
Logs a message with \code{\link{LOG_WARN}} severity.
}
