% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{set_log_level}
\alias{set_log_level}
\title{Sets the current log level (increasing order of severity: \code{\link{LOG_DEBUG}}, \code{\link{LOG_INFO}}, \code{\link{LOG_WARN}}, \code{\link{LOG_ERROR}}).}
\usage{
set_log_level(log_level)
}
\arguments{
\item{log_level}{The log level to set}
}
\description{
Sets the current log level (increasing order of severity: \code{\link{LOG_DEBUG}}, \code{\link{LOG_INFO}}, \code{\link{LOG_WARN}}, \code{\link{LOG_ERROR}}).
}
